+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [v0.2.5] - 2020-04-10
 ### Added
 - Add `SpecPredicate` and `tag_maybe` to the public interface (#49)
 
@@ -72,6 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial commit
 
+
+[v0.2.5]: https://github.com/coverahealth/dataspec/compare/v0.2.4..v0.2.5
 [v0.2.4]: https://github.com/coverahealth/dataspec/compare/v0.2.3..v0.2.4
 [v0.2.3]: https://github.com/coverahealth/dataspec/compare/v0.2.2..v0.2.3
 [v0.2.2]: https://github.com/coverahealth/dataspec/compare/v0.2.1..v0.2.2